requirements.txt
tqdm
streamlit>=0.89
pandas>=1.1.4
matplotlib>=3.3.3
numpy>=1.19.5
pytest>=6.2.3
Cerberus>=1.3.3
CoolProp>=6.4.1
PyYAML>=5.4.1
scipy>=1.6.0
